Install gems using phase 1 ruby properly
Following @74d3792 or one of the subsequent commits the build
directories were moved but the path was not updated.  This meant we used
whichever other ruby was in the path instead of the one we just built.

Also, now we use environment variables to pick the correct location gem
installation instead of --install-dir

Also, now we ensure that gem executables get installed in the correct
directory for use in the build process.